base --> MRTextInput
