+++ src/core/entities/MRMedia.js
