+++ src/core/entities/MRModel.js
@@ -1,12 +1,12 @@
 import { MREntity } from 'mrjs/core/MREntity';
 
 import { mrjsUtils } from 'mrjs';
-import { MRDivEntity } from '../MRDivEntity';
+import { MRDivEntity } from 'mrjs/core/MRDivEntity';
 
 /**
  * @class MRModel
  * @classdesc Loads in any supported 3D model type to the requested location. `mr-model`
- * @augments MREntity
+ * @augments MRDivEntity
  */
 export class MRModel extends MRDivEntity {
     /**
